[Fix] schedule lambda and one-day scraping lambda for pipeline backup
