+++ functions/lambda/scheduling/dormitory.py
@@ -2,8 +2,6 @@ import asyncio
 import json
 import logging
 
-from functions.shared.models.model import RestaurantType
-
 logger = logging.getLogger(__name__)
 
 
@@ -21,14 +19,12 @@ def dormitory_schedule_view(event, context):
         from functions.shared.utils.date_utils import get_next_weekdays, get_current_weekdays
         weekdays = get_current_weekdays() if delayed_schedule else get_next_weekdays()
 
-        # 3. 공통 스케줄링 서비스 사용
+        # 3. 기숙사 전용 스케줄링 서비스 사용
         from functions.config.dependencies import get_container
         container = get_container()
         scheduling_service = container.get_scheduling_service()
 
-        results = asyncio.run(scheduling_service.process_weekly_schedule(
-            RestaurantType.DORMITORY, weekdays
-        ))
+        results = asyncio.run(scheduling_service.process_weekly_schedule_dormitory(weekdays))
 
         logger.info("기숙사식당 주간 스케줄 완료")
