+++ functions/lambda/scheduling/faculty.py
@@ -7,30 +7,30 @@ from functions.shared.models.model import RestaurantType
 logger = logging.getLogger(__name__)
 
 
-def haksik_schedule_view(event, context):
-    """학생식당 주간 스케줄 뷰"""
+def faculty_schedule_view(event, context):
+    """교직원식당 주간 스케줄 뷰"""
     try:
         # 1. 파라미터 추출
         delayed_schedule = False
         if event.get("queryStringParameters"):
             delayed_schedule = bool(event["queryStringParameters"].get("delayed_schedule"))
 
-        logger.info(f"학생식당 주간 스케줄 시작 (delayed: {delayed_schedule})")
+        logger.info(f"교직원식당 주간 스케줄 시작 (delayed: {delayed_schedule})")
 
         # 2. 날짜 목록 결정
         from functions.shared.utils.date_utils import get_next_weekdays, get_current_weekdays
         weekdays = get_current_weekdays() if delayed_schedule else get_next_weekdays()
 
-        # 3. 공통 스케줄링 서비스 사용
+        # 3. 일반 식당용 스케줄링 서비스 사용
         from functions.config.dependencies import get_container
         container = get_container()
         scheduling_service = container.get_scheduling_service()
 
-        results = asyncio.run(scheduling_service.process_weekly_schedule(
-            RestaurantType.HAKSIK, weekdays
+        results = asyncio.run(scheduling_service.process_weekly_schedule_general(
+            RestaurantType.FACULTY, weekdays
         ))
 
-        logger.info("학생식당 주간 스케줄 완료")
+        logger.info("교직원식당 주간 스케줄 완료")
 
         return {
             'statusCode': 200,
@@ -39,17 +39,17 @@ def haksik_schedule_view(event, context):
         }
 
     except Exception as e:
-        logger.error(f"학생식당 스케줄링 오류: {e}", exc_info=True)
+        logger.error(f"교직원식당 스케줄링 오류: {e}", exc_info=True)
 
         return {
             'statusCode': 500,
             'headers': {'Content-Type': 'application/json; charset=utf-8'},
             'body': json.dumps({
                 'error': 'Internal server error',
-                'restaurant': 'HAKSIK'
+                'restaurant': 'FACULTY'
             }, ensure_ascii=False)
         }
 
 
 # AWS Lambda 핸들러
-lambda_handler = haksik_schedule_view
+lambda_handler = faculty_schedule_view
