+++ requirements.txt
@@ -1,0 +1,44 @@
+aiohappyeyeballs==2.6.1
+aiohttp==3.12.14
+aiosignal==1.4.0
+annotated-types==0.7.0
+anyio==4.9.0
+attrs==25.3.0
+beautifulsoup4==4.13.4
+bs4==0.0.2
+certifi==2025.7.14
+charset-normalizer==3.4.2
+coverage==7.10.0
+distro==1.9.0
+execnet==2.1.1
+frozenlist==1.7.0
+h11==0.16.0
+httpcore==1.0.9
+httpx==0.28.1
+idna==3.10
+iniconfig==2.1.0
+jiter==0.10.0
+multidict==6.6.3
+openai==1.95.1
+packaging==25.0
+pluggy==1.6.0
+propcache==0.3.2
+psutil==7.0.0
+pydantic==2.11.7
+pydantic-core==2.33.2
+pygments==2.19.2
+pytest==8.4.1
+pytest-asyncio==1.1.0
+pytest-cov==6.2.1
+pytest-mock==3.14.1
+pytest-xdist==3.8.0
+pytz==2025.2
+requests==2.32.4
+sniffio==1.3.1
+soupsieve==2.7
+tenacity==9.1.2
+tqdm==4.67.1
+typing-extensions==4.14.1
+typing-inspection==0.4.1
+urllib3==2.5.0
+yarl==1.20.1
